Move `BoxedUint` under `uint::boxed`
This layout maps better to adding `boxed` support to other submodules,
like `modular::boxed` (which could contain a `BoxedResidue`)
