+++ src/uint/boxed/encoding.rs
